refactored lazy updates
